requirements.txt
streamlit
google-generativeai
newsapi-python
moviepy
python-dotenv